image: enable smt
Disabling SMT inside the image creates problems on AWS. If we need
nosmt there is a cpu option available on AWS.
SMT should not affect AMD CPUs. Enabling SMT could be a problem if we
assume trusted and untrusted code inside the same CVM. But this is not
our threat model.